requirements.txt
aiofiles==23.1.0
html5tagger==1.3.0
httptools==0.5.0
multidict==6.0.4
neo4j-driver==4.4.10
neobolt==1.7.17
neomodel==5.0.1
python-dotenv==1.0.0
pytz==2023.3
sanic==23.3.0
sanic-routing==22.8.0
six==1.16.0
tracerite==1.1.0
ujson==5.8.0
uvloop==0.17.0
websockets==11.0.3
